requirements.txt
pyqt5==5.15.10
pillow==10.0.0
img2pdf==0.4.4
pytest==7.4.0
